consolidate DrillDownFiltersPanel stories
